Added routing
to fix/complete routing + to add dynamic name
