+++ basic-cra-login/src/Hello.jsx
@@ -1,11 +1,11 @@
-import { name } from './basic-cra-login/src/App';
-
 export const Hello = () => {
   return (
     <div>
-      <p>Hello {name}</p>
+      <p>Hello /INSERT NAME/</p>
     </div>
   );
 };
+
+//  Learn to use React context
 
 export default Hello;
